+++ tests/testthat.R
@@ -1,3 +1,7 @@
+if (!requireNamespace("testthat", quietly = TRUE)) {
+  quit(save = "no", status = 0)
+}
+
 library(testthat)
 library(inferMM)